+++ R/visualize_pathways_year_level1.R
@@ -44,7 +44,7 @@
 #' @param first_observed character. Name of the column of `df` containing
 #'  information about year of introduction. Default: `"first_observed"`.
 #' @param cbd_standard logical. If `TRUE` the values of pathway level 1 are
-#'   checked based on CBD standard as returned by `pathways_cbd()`. Error is
+#'   checked based on CBD standard as in `pathwayscbd`. Error is
 #'   returned if unmatched values are found. If `FALSE`, a warning is returned.
 #'   Default: `TRUE.`
 #' @param title `NULL` or character. Title of the graph. Default: `NULL`.
@@ -59,6 +59,7 @@
 #' - `data_facet_graph`: data.frame (tibble) with data used for the faceting
 #' plot in `plot`. `NULL` is returned if `facet_column` is `NULL`.
 #'
+#' @family checklist functions
 #' @export
 #' @importFrom dplyr %>% .data
 #' @importFrom rlang !!
@@ -341,7 +342,7 @@ visualize_pathways_year_level1 <- function(
     ))
   # Import all CBD pathways level 1
   pathways_level1_all <-
-    pathways_cbd() %>%
+    pathwayscbd %>%
     dplyr::distinct(.data$pathway_level1)
   # Select pathways
   if (!is.null(pathways)) {