+++ R/visualize_pathways_year_level2.R
@@ -48,7 +48,7 @@
 #' @param first_observed character. Name of the column of `df` containing
 #'   information about year of introduction. Default: `"first_observed"`.
 #' @param cbd_standard logical. If `TRUE` the values of pathway level 1 are
-#'   checked based on CBD standard as returned by `pathways_cbd()`. Error is
+#'   checked based on CBD standard as in `pathwayscbd`. Error is
 #'   returned if unmatched values are found. If `FALSE`, a warning is returned.
 #'   Default: `TRUE`.
 #' @param title `NULL` or character. Title of the graph. Default: `NULL`.
@@ -63,6 +63,7 @@
 #' - `data_facet_graph`: data.frame (tibble) with data used for the faceting
 #' plot in `plot`. `NULL` is returned if `facet_column` is `NULL`.
 #'
+#' @family checklist functions
 #' @export
 #' @importFrom dplyr %>% .data
 #' @importFrom rlang !!
@@ -405,7 +406,7 @@ visualize_pathways_year_level2 <- function(
     ))
   # Import all CBD pathways level 2 within chosen pathway level 1
   pathways_level2_all <-
-    pathways_cbd() %>%
+    pathwayscbd %>%
     dplyr::filter(.data$pathway_level1 == chosen_pathway_level1) %>%
     dplyr::distinct(.data$pathway_level2)
   # Select pathways