+++ NEWS.md
@@ -1,3 +1,8 @@
+# trias 3.1.3
+
+- Function `pathways_cbd()` is deprecated. Use data frame `pathwayscbd` directly (#134).
+- Add vignettes about checklist and occurrence based functions (#170).
+
 # trias 3.1.2
 
 - Allow empty pathways in `visualize_pathways_level1()` and `visualize_pathways_level2()` functions (#168).